refactor(Client engines): refactor code structure to access to each engine easier #53
