+++ src/client/RenderEngine/RenderingEngine.hpp
@@ -1,0 +1,29 @@
+/*
+** EPITECH PROJECT, 2024
+** R-Type-epitech
+** File description:
+** RenderingEngine
+*/
+
+#ifndef RENDERINGENGINE_HPP_
+#define RENDERINGENGINE_HPP_
+
+#include "../AEngine.hpp"
+#include "../Mediator/IMediator.hpp"
+
+namespace RType {
+    class RenderingEngine : public AEngine {
+        public:
+            RenderingEngine();
+            ~RenderingEngine();
+
+            /**
+             * @brief Run current instance of the engine.
+             */
+            void run() override;
+        protected:
+        private:
+    };
+}
+
+#endif /* !RENDERINGENGINE_HPP_ */
